requirements.txt
Django==1.10.7
django-taggit==0.22.1
olefile==0.44
Pillow==4.1.1
pytz==2017.2
